+++ web/src/components/ui/index.ts
@@ -526,3 +526,4 @@ export * from './chart-line-dynamic-rsi';
 export * from './chart-line-cum-tick';
 export * from './chart-line-fractal-bb';
 export * from './chart-line-recombining-kc';
+export * from './chart-line-bb-fisher';